extracting out events to their own module
