+++ bitacross-worker/core-primitives/types/src/parentchain/events.rs
@@ -1,0 +1,69 @@
+use crate::ShardIdentifier;
+use alloc::format;
+use codec::{Decode, Encode};
+use core::fmt::Debug;
+use itp_utils::stringify::account_id_to_string;
+use substrate_api_client::ac_node_api::StaticEvent;
+
+#[derive(Encode, Decode, Debug)]
+pub struct ExtrinsicSuccess;
+
+impl StaticEvent for ExtrinsicSuccess {
+	const PALLET: &'static str = "System";
+	const EVENT: &'static str = "ExtrinsicSuccess";
+}
+
+#[derive(Encode, Decode)]
+pub struct ExtrinsicFailed;
+
+impl StaticEvent for ExtrinsicFailed {
+	const PALLET: &'static str = "System";
+	const EVENT: &'static str = "ExtrinsicFailed";
+}
+
+#[derive(Encode, Decode, Debug)]
+pub struct BalanceTransfer {
+	pub from: AccountId,
+	pub to: AccountId,
+	pub amount: Balance,
+}
+
+impl core::fmt::Display for BalanceTransfer {
+	fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
+		let message = format!(
+			"BalanceTransfer :: from: {}, to: {}, amount: {}",
+			account_id_to_string::<AccountId>(&self.from),
+			account_id_to_string::<AccountId>(&self.to),
+			self.amount
+		);
+		write!(f, "{}", message)
+	}
+}
+
+impl StaticEvent for BalanceTransfer {
+	const PALLET: &'static str = "Balances";
+	const EVENT: &'static str = "Transfer";
+}
+
+#[derive(Encode, Decode, Debug)]
+pub struct ParentchainBlockProcessed {
+	pub shard: ShardIdentifier,
+	pub block_number: BlockNumber,
+	pub block_hash: Hash,
+	pub task_merkle_root: Hash,
+}
+
+impl core::fmt::Display for crate::parentchain::ParentchainBlockProcessed {
+	fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
+		let message = format!(
+			"ParentchainBlockProcessed :: nr {} shard: {}, merkle: {:?}, block hash {:?}",
+			self.block_number, self.shard, self.task_merkle_root, self.block_hash
+		);
+		write!(f, "{}", message)
+	}
+}
+
+impl StaticEvent for ParentchainBlockProcessed {
+	const PALLET: &'static str = "Teebag";
+	const EVENT: &'static str = "ParentchainBlockProcessed";
+}
